requirements.txt
colorama==0.4.4
pycryptodome==3.16.0
